devices: refactor IOAPIC to support multiple hypervisor backends
This commit refactors the IOAPIC emulation logic to decouple common register state management from hypervisor-specific interrupt injection. This is a prerequisite for supporting the Windows Hypervisor Platform (WHP).

It introduces the IoApicBackend trait to handle backend-specific logic such as interrupt routing updates and injection.
It moves the register state into a dedicated IoApicRegs struct, making the core IoApic struct generic over the backend (IoApic<B: IoApicBackend>) and place KVM-sspecific implementation into a new ioapic_kvm.rs file.
